Added OS folder and fixed DAA folder Merge sort in DAA was broken and converted into using only one array. Changed the os.c file into individual programs stored in OS folder.
