+++ OS/Lab_4.c
@@ -1,0 +1,29 @@
+#include<stdio.h>
+#include<stdlib.h>
+#include<unistd.h>
+#include<pthread.h>
+
+int g = 0;
+
+void *myThreadFun(void *vargp)
+{
+	int *myid = (int*)vargp; 
+	static int s = 0;
+	++s;
+    ++g;
+	printf("Thread ID : %u, static : %d, Global:%d\n", *myid, ++s, ++g);
+}
+
+int main()
+{
+	int i;
+	pthread_t tid;
+	
+    for(int i=0; i<3; i++)
+	{
+		pthread_create(&tid, NULL, myThreadFun, (void*)&tid);
+	}
+	pthread_exit(NULL);
+	
+    return 0;
+}
